requirements.txt
aniso8601==9.0.1
apispec==5.2.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
flasgger==0.9.5
Flask==2.1.3
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
jsonschema==4.9.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
mistune==2.0.4
npx==0.1.1
numpy==1.23.1
optional-django==0.3.0
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
react==4.3.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.11
Werkzeug==2.2.1
